prevent early roleover and handle roleover

+++ stack/modules/d7ap/phy.c
@@ -627,8 +627,8 @@ error_t phy_init(void) {
 }
 
 void status_write() {
-    uint16_t bg_trigger_ratio = 1024 * total_rssi_triggers / total_bg;
-    uint16_t scan_timeout_ratio = 1024 * (total_fg - total_succeeded_fg) / total_fg;
+    uint16_t bg_trigger_ratio = 1024 * (uint64_t)total_rssi_triggers / total_bg;
+    uint16_t scan_timeout_ratio = 1024 * (uint64_t)(total_fg - total_succeeded_fg) / total_fg;
     uint8_t buffer[4] = {(uint8_t)(bg_trigger_ratio >> 8), (uint8_t)(bg_trigger_ratio & 0xFF), (uint8_t)(scan_timeout_ratio >> 8), (uint8_t)(scan_timeout_ratio & 0xFF)};
     d7ap_fs_write_file(D7A_FILE_DLL_STATUS_FILE_ID, 8, buffer, 4);
 }
